Few changes in db and some modification of entity classes

+++ Software/STONKS/EntitiesLayer/Entities/StavkaPrimke.cs
@@ -17,7 +17,7 @@ namespace EntitiesLayer.Entities
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
         public int primka_id { get; set; }
 
-        public int rabat { get; set; }
+        public double rabat { get; set; }
 
         public int kolicina { get; set; }
